token tags refactor
