+++ README.md
@@ -1,5 +1,5 @@
 
-# Рекурсинвый спуск
+# Рекурсивный спуск
 
 1. Написать парсер методом рекурсивного спуска.
