+++ parser.vcxproj.filters
@@ -28,9 +28,6 @@
     <Filter Include="Header Files\Parser\Lexeme\Token\Enum">
       <UniqueIdentifier>{6c72bbf2-25b7-4105-9b04-a756579161bb}</UniqueIdentifier>
     </Filter>
-    <Filter Include="Header Files\Parser\Lexeme\Token\Iidentifier">
-      <UniqueIdentifier>{bac75629-3877-4b8c-b6a7-be84f3e5146f}</UniqueIdentifier>
-    </Filter>
     <Filter Include="Header Files\Parser\Lexeme\Token\Delimeter">
       <UniqueIdentifier>{be90c461-084e-47dc-a2fc-f622750abf1f}</UniqueIdentifier>
     </Filter>
@@ -64,6 +61,9 @@
     <Filter Include="Header Files\Common\Types">
       <UniqueIdentifier>{32ab75bc-56be-432b-813e-10dd2a50bd3f}</UniqueIdentifier>
     </Filter>
+    <Filter Include="Header Files\Parser\Lexeme\Token\Special">
+      <UniqueIdentifier>{bac75629-3877-4b8c-b6a7-be84f3e5146f}</UniqueIdentifier>
+    </Filter>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="src\main.cpp">
@@ -95,25 +95,25 @@
     <ClInclude Include="src\parser\lexeme\CLexeme\CLexeme.hpp">
       <Filter>Header Files\Parser\Lexeme\CLexeme</Filter>
     </ClInclude>
-    <ClInclude Include="src\parser\lexeme\token\all.h">
+    <ClInclude Include="src\parser\lexeme\token\definitions\token_definitions.h">
       <Filter>Header Files\Parser\Lexeme\Token</Filter>
     </ClInclude>
-    <ClInclude Include="src\parser\lexeme\token\BaseToken.h">
+    <ClInclude Include="src\parser\lexeme\token\definitions\BaseToken.h">
       <Filter>Header Files\Parser\Lexeme\Token</Filter>
     </ClInclude>
-    <ClInclude Include="src\parser\lexeme\token\delimeter\TokenDelimeter.h">
+    <ClInclude Include="src\parser\lexeme\token\definitions\delimeter\TokenDelimeter.h">
       <Filter>Header Files\Parser\Lexeme\Token\Delimeter</Filter>
     </ClInclude>
-    <ClInclude Include="src\parser\lexeme\token\enum\TokenEnum.hpp">
+    <ClInclude Include="src\parser\lexeme\token\definitions\enum\TokenEnum.hpp">
       <Filter>Header Files\Parser\Lexeme\Token\Enum</Filter>
     </ClInclude>
-    <ClInclude Include="src\parser\lexeme\token\identifier\TokenIdentifier.h">
-      <Filter>Header Files\Parser\Lexeme\Token\Iidentifier</Filter>
+    <ClInclude Include="src\parser\lexeme\token\identifier\TokenSpecial.h">
+      <Filter>Header Files\Parser\Lexeme\Token\Special</Filter>
     </ClInclude>
-    <ClInclude Include="src\parser\lexeme\token\keyword\TokenKeyword.h">
+    <ClInclude Include="src\parser\lexeme\token\definitions\keyword\TokenKeyword.h">
       <Filter>Header Files\Parser\Lexeme\Token\Keyword</Filter>
     </ClInclude>
-    <ClInclude Include="src\parser\lexeme\token\type\TokenType.h">
+    <ClInclude Include="src\parser\lexeme\token\definitions\type\TokenType.h">
       <Filter>Header Files\Parser\Lexeme\Token\Type</Filter>
     </ClInclude>
     <ClInclude Include="src\parser\lexeme\exception\ILexemeException.h">
@@ -138,6 +138,12 @@
       <Filter>Header Files\Common\Types</Filter>
     </ClInclude>
     <ClInclude Include="src\parser\Parser.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="src\parser\lexeme\token\definitions\op\TokenOperator.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="src\parser\lexeme\token\definitions\special\TokenSpecial.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
