+++ src/parser/lexeme/token/definitions/op/TokenOperator.h
@@ -1,0 +1,15 @@
+#pragma once
+
+#include "../BaseToken.h"
+
+namespace parser::lexeme::token::op
+{
+
+// clang-format off
+
+struct TokenDelimeter : Token {};
+
+struct AssignmentOp       : TokenDelimeter {}; // :=
+struct PlusOp             : TokenDelimeter {}; // +
+struct MultiplyOp         : TokenDelimeter {}; // *
+}
