--- src/parser/lexeme/token/identifier/TokenIdentifier.h
@@ -1,17 +1,0 @@
-#pragma once
-
-#include "../BaseToken.h"
-
-namespace parser::lexeme::token::identifier
-{
-
-// clang-format off
-
-struct TokenIdentificator : Token {};
-
-struct TokenIdentificatorId  : TokenIdentificator {};
-struct TokenIdentificatorNum : TokenIdentificator {};
-
-// clang-format on
-
-} // namespace parser::lexeme::token::identifier
